zerotier-one 1.0.2 try #2

+++ zerotier-one-api/src/models/controller_status.rs
@@ -18,7 +18,7 @@ pub struct ControllerStatus {
     #[serde(rename = "apiVersion", skip_serializing_if = "Option::is_none")]
     pub api_version: Option<i32>,
     #[serde(rename = "clock", skip_serializing_if = "Option::is_none")]
-    pub clock: Option<i32>,
+    pub clock: Option<i64>,
 }
 
 impl ControllerStatus {
